fix: SendGrid email TemplateData should be Dictionary for now
Before it was an `object` but that led to it coming through as `JsonElement` when being deserialized from the integration event payload. It caused some emails to come through as if the template data was not specified. SendGrid templates can use `{{{Subject}}}` variables, which did not work at all before.